Implemented modal window functionality by opening a new window of information
Functionality includes: Opening a window with information and blurring the the area outside that window. The 'X' button closes the window and returns the website back to normal. Simply added a function to traverse through the modal buttons and called a function to listen for click events to add or remove hidden from the class.